Remove loggers, rename Executor interface to ShellExecutor, remove copy method from ShellExecutor interface, rename ShellExecutor implementation to LocalShellExecutor
