+++ osbenchmark/builder/executors/exception_handling_shell_executor.py
@@ -1,0 +1,13 @@
+from osbenchmark.builder.executors.shell_executor import ShellExecutor
+from osbenchmark.exceptions import ExecutorError
+
+
+class ExceptionHandlingShellExecutor(ShellExecutor):
+    def __init__(self, executor):
+        self.executor = executor
+
+    def execute(self, host, command, **kwargs):
+        try:
+            self.executor.execute(host, command, kwargs)
+        except Exception as e:
+            raise ExecutorError("Command \"{}\" on host \"{}\" failed to execute".format(command, host), e)
