+++ osbenchmark/builder/executors/shell_executor.py
@@ -1,19 +1,15 @@
-import shutil
-import subprocess
+class ShellExecutor:
+    """
+    Executors are used to run shell commands on the cluster hosts. Implementations of this class will use various
+    technologies to interface with the hosts of a cluster.
+    """
 
-from osbenchmark.builder.executors.executor import Executor
-from osbenchmark.exceptions import ExecutorError
-from osbenchmark.utils import process
+    def execute(self, host, command, **kwargs):
+        """
+        Executes a list of commands against the provided host
 
-
-class ShellExecutor(Executor):
-    # pylint: disable=arguments-differ
-    def execute(self, host, command, output=False, stdout=subprocess.PIPE, stderr=subprocess.PIPE, env=None, detach=False):
-        if output:
-            return process.run_subprocess_with_output(command)
-        else:
-            if process.run_subprocess_with_logging(command, stdout=stdout, stderr=stderr, env=env, detach=detach):
-                raise ExecutorError("Command: \"{}\" returned a non-zero exit code".format(command))
-
-    def copy(self, host, source, destination):
-        shutil.copy(source, destination)
+        ;param host: A Host object defining the host on which to execute the commands
+        ;param command: A shell command as a string
+        ;return output: The output of the command
+        """
+        raise NotImplementedError
